create-parser-functions: Separate files.
